PDO-629/631 Remove PA Admin Restart and Refactoring
PDO-629: PingAccess Admin node was requiring a restart in order to pick up the latest Admin KeyPairs. Jira ticket PA-11748 addressed this issue but by default the restart was required. This commit specifies to turn off restart required allowing some code to be removed.

PDO-631: Wait functions were added due to a race condition between when PA admin was marked as ready but still finishing configuration and when PA engine was started. This commit removes the additional wait functions and adds an additional requirement within the PA engine init container that the configuration must be complete on the admin node before the engine node can startup.

Code Review: Add a check that will kill the container if the background process started in 80-post-start.sh has any configuration steps that fail.