[skip pipeline] Include ping-cloud-base commit SHA that was used to set up the initial cluster state repo
This is useful when updating from one version to another, especially when the version is referenced by a sliding branch as opposed to an immutable tag. For example, lab and preview US1 will always use the sliding branches ref master and the current release branch (e.g. v1.3.0-release-branch). When updating the cluster state repo with new changes that have since been pushed into ping-cloud-base, it is useful to know the commit SHA so exact diffs can be copied over into the k8s-configs, especially.