PDO-846: prevent engine flapping due to flux/HPA interfering with each other
When "replicas" are explicitly set in the Deployment/StatefulSet object referenced by an HPA, Flux (the Continuous Delivery) will apply that, which could interfere with auto-scaling previously performed by the HPA. While we were careful not to set the "replicas" in the base yaml files, they were being set in downstream kustomization.yaml files being used in PingCloud. This caused the PingAccess engine nodes to flap in various environments.