PDO-928: workaround for DS-41964: use --replaceFullProfile to replace-profile
The optimization that was added in 8.0.0.1 to skip doing a full replace unless there are only configuration changes in the old and new profiles caused this regression. The internal tool that PD uses (config-diff) wasn't figuring changing the environment variables as a config change and nothing else had changed in the profile. So replace-profile ended up being a no-op. So the changes to valid P14C secrets didn't stick.