PDO-1397: add a script to update the cluster-state repo from one release to another
When resetting to default, do not source the environment variables from the default
CDE branch.

Also, always create the *.old files for env_vars, secrets.yaml and sealed-secrets.yaml
to make it easier for the operator to see the differences between the new branch and
its default CDE branch.

Changing flux-git-deploy to argo-git-deploy in v1.8-release-branch.