PCPTSD-823: use a dir under /opt/out  as the staging dir for restore
We'd been using a directory under the /tmp as the staging directory for
restore. The tmp file system is the host file system and may not have enough
space to hold the backup file. In that case, restores of backup will fail.
Instead of using the tmp file system, this MR changes the restore staging
directory to be under the /opt/out volume, which is an EBS volume in
Ping Cloud and has more disk capacity.