PDO-2067: fixing seal.sh to run flux-command.sh with a region directory
Previously, flux-command.sh was being invoked by pointing it to the 'base'
directory, and fake values were supplied to it for environment variables
defined in regional directories. When IRSA variable was added, a fake value
was missed for it, which broke seal.sh. The implementation was brittle and
would need fixing every time a new variable was added in the regional env_vars.
So to fix it, now we're pointing flux-command.sh to a region directory - it
can be any one of them for it to work.